+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+Unreleased
+----------
+
+New features
+************
+
+* Add docx_pagebreak_before_table_of_contents option.
+
 Release 1.1.4 (2019-09-25)
 --------------------------